Update PageViewStatDTO to use double for count properties

+++ dashboard/backend/Application/DTOs/AnalyticsData/PageViewStatDTO.cs
@@ -3,9 +3,9 @@ namespace Application.DTOs.AnalyticsData
     public class PageViewStatDTO
     {
         public required string Url { get; set; }
-        public required int LandingCount { get; set; }
-        public required int ExitCount { get; set; }
-        public required int BounceCount { get; set; }
+        public required double LandingCount { get; set; }
+        public required double ExitCount { get; set; }
+        public required double BounceCount { get; set; }
         public required int Count { get; set; }
         public required double AvgInteractionCount { get; set; }
     }
